# doc/api.yml
openapi: 3.0.1
info:
  title: PROG4 Final Exam - School Management API
  description: |
    REST API of the school-management platform (Poja/Spring Boot 3.2.2, base package
    `org.cocojojo.mg`).

    Authentication is JWT-based. Obtain a token via `POST /auth/login`, then send it as
    `Authorization: Bearer <token>`. Three roles exist: `ADMIN`, `TEACHER`, `STUDENT`.
    Role restrictions are declared per operation.

    Error responses use a single shape:
    ```json
    { "error": "<message>", "status": "BAD_REQUEST" }
    ```
    with status codes 400, 401, 403, 404, 409 (see the `Error` schema and common responses).
  version: v1
  contact:
    name: PROG4 team
servers:
  - url: https://qcbjel2dhzmk43tv65e6vuqe4e0fhsaj.lambda-url.eu-west-3.on.aws
    description: Production (deployed Poja lambda)
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /exams/{examId}/grades:
    get:
      tags: [Grades]
      summary: List grades of an exam
      description: Requires the ADMIN or TEACHER role.
      operationId: listExamGrades
      parameters:
        - $ref: '#/components/parameters/ExamIdPath'
      responses:
        '200':
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Grades]
      summary: Enter grades for an exam (bulk)
      description: |
        Grades a whole class in a single request (one grade per student). The batch is validated as
        a unit before any row is inserted. Requires the ADMIN or TEACHER role.
      operationId: createExamGrades
      parameters:
        - $ref: '#/components/parameters/ExamIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/GradeRequest'
      responses:
        '200':
          description: Grades created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /exams/{examId}/students/{studentId}/grade:
    get:
      tags: [Grades]
      summary: Get a student's grade for an exam
      description: Requires the ADMIN or TEACHER role.
      operationId: getExamGradeForStudent
      parameters:
        - $ref: '#/components/parameters/ExamIdPath'
        - $ref: '#/components/parameters/StudentIdPath'
      responses:
        '200':
          description: Grade found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Grades]
      summary: Correct a grade
      description: |
        Overrides a grade and records the correction (previous value, new value, reason, author) in
        the grade history. Returns the full correction history (most recent first). Requires the
        ADMIN or TEACHER role.
      operationId: correctGrade
      parameters:
        - $ref: '#/components/parameters/ExamIdPath'
        - $ref: '#/components/parameters/StudentIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeCorrectionRequest'
      responses:
        '200':
          description: Grade corrected, full correction history returned (most recent first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeHistoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /students/{studentId}/grades:
    get:
      tags: [Grades]
      summary: List a student's grades
      description: Requires the ADMIN, TEACHER or STUDENT role.
      operationId: listStudentGrades
      parameters:
        - $ref: '#/components/parameters/StudentIdPath'
      responses:
        '200':
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /grades/{gradeId}:
    get:
      tags: [Grades]
      summary: Get a grade by id
      description: Requires the ADMIN, TEACHER or STUDENT role.
      operationId: getGrade
      parameters:
        - $ref: '#/components/parameters/GradeIdPath'
      responses:
        '200':
          description: Grade found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Grades]
      summary: Delete a grade
      description: |
        Deletes a grade; the deletion reason is required and recorded for traceability. Returns the
        full history including the deletion record (most recent first). Requires the ADMIN or TEACHER
        role.
      operationId: deleteGrade
      parameters:
        - $ref: '#/components/parameters/GradeIdPath'
        - name: reason
          in: query
          required: true
          description: Reason for deleting the grade, recorded for traceability.
          schema:
            type: string
      responses:
        '200':
          description: Grade deleted, full history returned (including the deletion record, most recent first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeHistoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /grades/{gradeId}/history:
    get:
      tags: [Grades]
      summary: Get a grade's correction history
      description: Returns every recorded correction of a grade. Requires the ADMIN, TEACHER or STUDENT role.
      operationId: getGradeHistory
      parameters:
        - $ref: '#/components/parameters/GradeIdPath'
      responses:
        '200':
          description: List of corrections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT obtained from `POST /auth/login` and sent as `Authorization: Bearer <token>`. The token
        is issued for an ADMIN, TEACHER or STUDENT account.
  parameters:
...
    ExamIdPath:
      name: examId
      in: path
      description: Exam UUID
      required: true
      schema:
        type: string
        format: uuid
    StudentIdPath:
      name: studentId
      in: path
      description: Student UUID
      required: true
      schema:
        type: string
        format: uuid
    GradeIdPath:
      name: gradeId
      in: path
      description: Grade UUID
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request (validation failure, bad coefficient, illegal argument or integrity violation)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing, invalid or expired token; bad credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        WWW-Authenticate:
          description: Present when no bearer token was supplied
          schema:
            type: string
    Forbidden:
      description: Authenticated but the role does not allow this operation, or caller is not the target user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: A uniqueness or business-rule conflict occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: Uniform error body produced by the global exception handler
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Grade with id 123 not found"
        status:
          type: string
          description: HTTP status phrase
          example: "NOT_FOUND"
...
    GradeRequest:
      type: object
      description: One grade entry of a bulk grading request
      required: [studentId, value]
      properties:
        studentId:
          type: string
          format: uuid
        value:
          type: number
          minimum: 0
          maximum: 20
          example: 14.5
        comment:
          type: string
    GradeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        studentStd:
          type: string
        examId:
          type: string
          format: uuid
        examTitle:
          type: string
        courseCode:
          type: string
        value:
          type: number
          example: 14.5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GradeCorrectionRequest:
      type: object
      description: Corrects an existing grade and records the change in history
      required: [value, reason]
      properties:
        value:
          type: number
          minimum: 0
          maximum: 20
          example: 16
        reason:
          type: string
          minLength: 1
          description: Mandatory justification of the correction
...
    GradeHistoryResponse:
      type: object
      description: One recorded correction of a grade
      properties:
        id:
          type: string
          format: uuid
        gradeId:
          type: string
          format: uuid
        previousValue:
          type: number
        newValue:
          type: number
        reason:
          type: string
        changedById:
          type: string
          format: uuid
        changedByName:
          type: string
        changedAt:
          type: string
          format: date-time
...